Fix tests in preparation of TestNG upgrade
1. API overloads cause compilation failure, updated to use the
correct APIs.
2. Tests run order are changing and more parallelism. Cleanup
after creating some views, so the later queries does not fail.
3. Handle multi threaded initialization gracefully in Kudu Runner.

+++ presto-tests/src/main/java/com/facebook/presto/tests/AbstractTestAggregations.java
@@ -857,8 +857,8 @@ public abstract class AbstractTestAggregations
         List<MaterializedRow> actualRows = actual.getMaterializedRows();
         assertEquals(actualRows.size(), 1);
         assertTrue(Double.isNaN(((List<Double>) actualRows.get(0).getField(0)).get(0)));
-        assertEquals(((List<Double>) actualRows.get(0).getField(0)).get(1), 2.0);
-        assertEquals(((List<Double>) actualRows.get(0).getField(0)).get(2), 3.0);
+        assertEquals(((List<Double>) actualRows.get(0).getField(0)).get(1).doubleValue(), 2.0);
+        assertEquals(((List<Double>) actualRows.get(0).getField(0)).get(2).doubleValue(), 3.0);
     }
 
     @Test
